moved SqlDatabase class into its own file, removed unnecessary console logs, added index.ts to export handlers from common file, made some type def changes to preload script, added component return types on front end, and added an in progress login page
