+++ src/main/handlers/handleEncryptPassword.ts
@@ -1,12 +1,11 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import Cryptr from 'cryptr';
+import Cryptr from "cryptr";
 
 export async function handleEncryptPassword(
   _: Electron.IpcMainInvokeEvent,
-  data: string[]
+  data: string[],
 ) {
-
-  const [__, password, secretKey] = data
+  const [__, password, secretKey] = data;
 
   const cryptr = new Cryptr(secretKey);
   const encryptedString = cryptr.encrypt(password);
