+++ src/main/handlers/handleForkUtilityProcess.ts
@@ -2,14 +2,12 @@ import path from "path";
 import { utilityProcess, MessageChannelMain } from "electron";
 
 // Main process
-const handleForkUtilityProcess = (
+export const handleForkUtilityProcess = (
   event: Electron.IpcMainInvokeEvent,
   data: string[]
 ) => {
   // take data received from ipcRenderer and pass on to
   // child process to encrypt passwords
-  console.log("handleForkUtilityProcess is being invoked. Data: ", data);
-  console.log(__dirname)
   const { port1 } = new MessageChannelMain();
   const [username, password] = data;
   const child = utilityProcess.fork(
@@ -19,5 +17,3 @@ const handleForkUtilityProcess = (
   );
   child.postMessage({ data }, [port1]);
 };
-
-export default handleForkUtilityProcess;
