+++ src/main/handlers/index.ts
@@ -1,0 +1,3 @@
+export * from "./handleEncryptPassword";
+export * from "./handleFileOpen";
+export * from "./handleForkUtilityProcess";
